+++ core/xml-reader/src/main/java/org/marceloleite/mercado/xml/converters/CurrencyAmountXmlConverter.java
@@ -1,5 +1,7 @@
 package org.marceloleite.mercado.xml.converters;
 
+import java.math.BigDecimal;
+
 import org.marceloleite.mercado.commons.Currency;
 import org.marceloleite.mercado.data.CurrencyAmountData;
 import org.marceloleite.mercado.xml.structures.XmlCurrencyAmount;
@@ -9,14 +11,14 @@ public class CurrencyAmountXmlConverter implements XmlConverter<XmlCurrencyAmoun
 	@Override
 	public XmlCurrencyAmount convertToXml(CurrencyAmountData currencyAmount) {
 		Currency currency = currencyAmount.getCurrency();
-		double amount = currencyAmount.getAmount();
+		BigDecimal amount = currencyAmount.getAmount();
 		return new XmlCurrencyAmount(currency, amount);
 	}
 
 	@Override
 	public CurrencyAmountData convertToObject(XmlCurrencyAmount xmlCurrencyAmount) {
 		Currency currency = xmlCurrencyAmount.getCurrency();
-		double amount = xmlCurrencyAmount.getAmount();
+		BigDecimal amount = xmlCurrencyAmount.getAmount();
 		return new CurrencyAmountData(currency, amount);
 	}
