+++ core/xml-reader/src/main/java/org/marceloleite/mercado/xml/structures/XmlBalances.java
@@ -1,5 +1,6 @@
 package org.marceloleite.mercado.xml.structures;
 
+import java.math.BigDecimal;
 import java.util.EnumMap;
 
 import javax.xml.bind.annotation.XmlRootElement;
@@ -7,7 +8,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 import org.marceloleite.mercado.commons.Currency;
 
 @XmlRootElement(name = "balances")
-public class XmlBalances extends EnumMap<Currency, Double> {
+public class XmlBalances extends EnumMap<Currency, BigDecimal> {
 
 	/**
 	 *
